+++ persistent-qq/test/PersistentTestModels.hs
@@ -9,6 +9,7 @@
 {-# LANGUAGE TemplateHaskell #-}
 {-# LANGUAGE TypeFamilies #-}
 {-# LANGUAGE UndecidableInstances #-} -- FIXME
+{-# LANGUAGE DerivingStrategies #-}
 module PersistentTestModels where
 
 import Control.Monad.Reader