+++ persistent-mongoDB/ChangeLog.md
@@ -1,5 +1,13 @@
 # Changelog for persistent-mongoDB
 
+## 2.9.0.2
+
+* Compatibility with latest mongoDB [#1012](https://github.com/yesodweb/persistent/pull/1012)
+
+## 2.9.0.1
+
+* Compatibility with latest persistent-template for test suite [#1002](https://github.com/yesodweb/persistent/pull/1002/files)
+
 ## 2.9.0
 
 * Removed deprecated `entityToDocument`. Please use `recordToDocument` instead. [#894](https://github.com/yesodweb/persistent/pull/894)