+++ persistent-sqlite/ChangeLog.md
@@ -1,5 +1,13 @@
 # Changelog for persistent-sqlite
 
+## 2.10.5.2
+
+* Compatibility with latest persistent-template for test suite [#1002](https://github.com/yesodweb/persistent/pull/1002/files)
+
+## 2.10.5.1
+
+* a fix for template-haskell 2.16, GHC 8.10 alpha [#993](https://github.com/yesodweb/persistent/pull/993) @simonmichael
+
 ## 2.10.5
 
 * Foreign keys table constraints are correctly generated [#945](https://github.com/yesodweb/persistent/pull/945) @kderme